+++ modules/smartContracts/internalModules/Contract.js
