+++ modules/smartContracts/internalModules/TokensRegister.js
